+++ core-tests/src/test/java/org/mule/runtime/core/processor/chain/DefaultMessageProcessorChainTestCase.java
@@ -117,7 +117,6 @@ public class DefaultMessageProcessorChainTestCase extends AbstractReactiveProces
         {new ProactorProcessingStrategyFactory(), BLOCKING},
         {new WorkQueueProcessingStrategyFactory(), BLOCKING},
         {new SynchronousProcessingStrategyFactory(), BLOCKING},
-        {new SynchronousStreamPerEventProcessingStrategyFactory(), BLOCKING},
         {new LegacySynchronousProcessingStrategyFactory(), BLOCKING},
         {new LegacyDefaultFlowProcessingStrategyFactory(), BLOCKING},
         {new LegacyNonBlockingProcessingStrategyFactory(), BLOCKING},
@@ -128,7 +127,6 @@ public class DefaultMessageProcessorChainTestCase extends AbstractReactiveProces
         {new ProactorProcessingStrategyFactory(), NON_BLOCKING},
         {new WorkQueueProcessingStrategyFactory(), NON_BLOCKING},
         {new SynchronousProcessingStrategyFactory(), BLOCKING},
-        {new SynchronousStreamPerEventProcessingStrategyFactory(), NON_BLOCKING},
         {new LegacySynchronousProcessingStrategyFactory(), NON_BLOCKING},
         {new LegacyDefaultFlowProcessingStrategyFactory(), NON_BLOCKING},
         {new LegacyNonBlockingProcessingStrategyFactory(), NON_BLOCKING},
