+++ tests/performance/src/main/java/org/mule/FlowBenchmark.java
@@ -40,7 +40,6 @@ public class FlowBenchmark extends AbstractBenchmark {
 
   @Param({
       "org.mule.runtime.core.processor.strategy.LegacySynchronousProcessingStrategyFactory",
-      "org.mule.runtime.core.processor.strategy.SynchronousStreamPerEventProcessingStrategyFactory",
       "org.mule.runtime.core.processor.strategy.SynchronousProcessingStrategyFactory",
       "org.mule.runtime.core.processor.strategy.ReactorProcessingStrategyFactory",
       "org.mule.runtime.core.processor.strategy.MultiReactorProcessingStrategyFactory",
